feat: Continuação dos mappings

+++ EmergencyManagementSystem.SAMU.Entities/Entities/TeamMember.cs
@@ -5,8 +5,7 @@ namespace EmergencyManagementSystem.SAMU.Entities.Entities
     public class TeamMember
     {
         public Guid EmployeeGuid { get; set; }
-        //public virtual Employee Employee { get; set; }
-        public int VehicleTeamId { get; set; }
+        public long VehicleTeamId { get; set; }
         public virtual VehicleTeam VehicleTeam { get; set; }
 
         //verificar como isso ficará registrado na ocorrência. quem estava no veiculo durante a ocorrencia
